Move cli build functions to controller
